trivial: move kernel version check into common library code
This will be used by more than one plugin in an upcoming commit.
