+++ libfwupdplugin/fu-common.c
@@ -21,6 +21,10 @@
 #include <cpuid.h>
 #endif
 
+#ifdef HAVE_UTSNAME_H
+#include <sys/utsname.h>
+#endif
+
 #ifdef HAVE_LIBARCHIVE
 #include <archive_entry.h>
 #include <archive.h>
@@ -34,6 +38,7 @@
 #include "fwupd-error.h"
 
 #include "fu-common-private.h"
+#include "fu-common-version.h"
 #include "fu-firmware.h"
 #include "fu-volume-private.h"
 
@@ -2567,6 +2572,54 @@ fu_common_kernel_locked_down (void)
 #else
 	return FALSE;
 #endif
+}
+
+/**
+ * fu_common_check_kernel_version :
+ * @minimum_kernel: (not nullable): The minimum kernel version to check against
+ * @error: (nullable): optional return location for an error
+ *
+ * Determines if the system is running at least a certain required kernel version
+ *
+ * Since: 1.6.2
+ **/
+gboolean
+fu_common_check_kernel_version (const gchar *minimum_kernel, GError **error)
+{
+#ifdef HAVE_UTSNAME_H
+	struct utsname name_tmp;
+
+	g_return_val_if_fail (error == NULL || *error == NULL, FALSE);
+	g_return_val_if_fail (minimum_kernel == NULL, FALSE);
+
+	memset (&name_tmp, 0, sizeof(struct utsname));
+	if (uname (&name_tmp) < 0) {
+		g_set_error_literal (error,
+				     FWUPD_ERROR,
+				     FWUPD_ERROR_INTERNAL,
+				     "failed to read kernel version");
+		return FALSE;
+	}
+	if (fu_common_vercmp_full (name_tmp.release,
+				   minimum_kernel,
+				   FWUPD_VERSION_FORMAT_TRIPLET) < 0) {
+		g_set_error (error,
+			     FWUPD_ERROR,
+			     FWUPD_ERROR_INTERNAL,
+			     "kernel %s doesn't meet minimum %s",
+			     name_tmp.release, minimum_kernel);
+		return FALSE;
+	}
+
+	return TRUE;
+#else
+	g_set_error_literal (error,
+			     FWUPD_ERROR,
+			     FWUPD_ERROR_INTERNAL,
+			     "platform doesn't support checking for minimum Linux kernel");
+	return FALSE;
+#endif
+
 }
 
 /**
